+++ test/test_schemas.py
@@ -1,6 +1,8 @@
 import pytest
 from jsonschema import RefResolver
 
+import schemathesis
+
 
 @pytest.mark.parametrize("base_path", ("/v1", "/v1/"))
 def test_base_path_suffix(swagger_20, base_path):
@@ -46,3 +48,43 @@ def test_resolver_cache(swagger_20, mocker):
     assert "_resolver" in swagger_20.__dict__
     assert isinstance(swagger_20.resolver, RefResolver)
     assert spy.call_count == 1
+
+
+def test_resolving_multiple_files():
+    raw_schema = {
+        "swagger": "2.0",
+        "info": {"title": "Example API", "description": "An API to test Schemathesis", "version": "1.0.0"},
+        "host": "127.0.0.1:8888",
+        "basePath": "/api",
+        "schemes": ["http"],
+        "produces": ["application/json"],
+        "paths": {
+            "teapot": {
+                "post": {
+                    "parameters": [
+                        {
+                            "schema": {"$ref": "test/data/petstore_v2.yaml#/definitions/User"},
+                            "in": "body",
+                            "name": "user",
+                            "required": True,
+                        }
+                    ]
+                }
+            }
+        },
+    }
+    schema = schemathesis.from_dict(raw_schema)
+    assert schema["/api/teapot"]["post"].body == {
+        "type": "object",
+        "properties": {
+            "id": {"type": "integer", "format": "int64"},
+            "username": {"type": "string"},
+            "firstName": {"type": "string"},
+            "lastName": {"type": "string"},
+            "email": {"type": "string"},
+            "password": {"type": "string"},
+            "phone": {"type": "string"},
+            "userStatus": {"type": "integer", "format": "int32", "description": "User Status"},
+        },
+        "xml": {"name": "User"},
+    }